BUG/MINOR: http-ana: Reset request analysers on error when waiting for response
This bug was supposed to be fixed by the commit f6df2b4a ("BUG/MINOR: http-ana:
Reset request analysers on a response side error"). It is a backported patch
from the 2.2. But, while it is enough on the 2.2, it is not on 2.1 and
lower. For these versions, the error handling is not grouped at the end of
analysers. At many places, when we are waiting for a response, several errors
are immediately returned. For all of these, the same fix must be applied.

The patch was directly introduced on 2.1, there is no upstream commit ID for
this patch. It must be backported everywhere the commit f6df2b4a is.

(cherry picked from commit dd6f0b1a74fb1241d276484f3c4aced513a95b78)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>

[Cf: The legacy code was also fixed the same way. The patch must be backported
to lower versions, as far as 1.7. In 1.9, it will be applied on the mentionned
commit, but for others, only this one, for the legacy part, will be necessary. ]

(cherry picked from commit 8868ac34aeb0bd1cf902820eff3ea49dcafea1ca)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 0f15008cf49c7a29a11621d2a43afbfa75c3225a)
[wt: dropped HTX code]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 25bcf8d7401c3f2ee89ae60241f11c0d43413920)
Signed-off-by: Willy Tarreau <w@1wt.eu>