MINOR: version: Show uname output in display_version()
This patch adds the sysname, release, version and machine fields from
the uname results to the version output. It intentionally leaves out the
machine name, because it is usually not useful and users might not want to
expose their machine names for privacy reasons.

May be backported if it is considered useful for debugging.

(cherry picked from commit dfad6a41ad9f012671b703788dd679cf24eb8c5a)
Signed-off-by: Willy Tarreau <w@1wt.eu>