REGTEST: ssl: test the client certificate authentication
This reg-test tests the client auth feature of HAProxy for both the
backend and frontend section with a CRL list.

This reg-test uses 2 chained listeners because vtest does not handle the
SSL. Test the frontend client auth and the backend side at the same
time.

It sends 3 requests: one with a correct certificate, one with an expired
one and one which was revoked. The client then checks if we received the
right one with the right error.

Certificates, CA and CRL are expiring in 2050 so it should be fine for
the CI.

This test could be backported as far as HAProxy 1.6

(cherry picked from commit 2e0dbb74126115e91adcf584f671cd9d1952ddd9)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 5dc38adfdb08290706ca9e1b0d35d74d2198d359)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 1d15f68a633f8b28a97327234b66969ced204bae)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>